rename typo dynymic
