+++ src/dynamicContent/dynamicContent.domcomponents.js
