+++ src/dynamicContent/dynamicContent.events.js
