+++ src/events.js
@@ -1,4 +1,4 @@
-import DynamicContentEvents from './dynymicContent/dynamicContent.events';
+import DynamicContentEvents from './dynamicContent/dynamicContent.events';
 
 export default (editor) => {
   const dce = new DynamicContentEvents(editor);
